Fix some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/3905



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@16769 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/cintltst/usettest.c
@@ -245,8 +245,9 @@ static void expectContainment(const USet* set,
             int32_t stringLength = 0;
             char strCopy[64];
 
-            while (*p++ != '}') {}
-            stringLength = p - stringStart - 1;
+            while (*p++ != '}') {
+            }
+            stringLength = (int32_t)(p - stringStart - 1);
             strncpy(strCopy, stringStart, stringLength);
             strCopy[stringLength] = 0;
 
@@ -391,8 +392,9 @@ static void expectItems(const USet* set,
             int32_t stringLength = 0;
             char strCopy[64];
 
-            while (*p++ != '}') {}
-            stringLength = p - stringStart - 1;
+            while (*p++ != '}') {
+            }
+            stringLength = (int32_t)(p - stringStart - 1);
             strncpy(strCopy, stringStart, stringLength);
             strCopy[stringLength] = 0;
